Move other sources to src, add basic ts stuff
