+++ package.json
@@ -1,14 +1,14 @@
 {
   "name": "discord-mod-bot",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "lint": "eslint . --ext js,vue",
     "lint-fix": "eslint . --ext js,vue --fix",
     "build": "webpack",
     "build:prod": "NODE_ENV=production webpack",
-    "start": "node core",
-    "start:prod": "NODE_ENV=production node core"
+    "start": "ts-node src/index.js",
+    "start:prod": "NODE_ENV=production ts-node src/index.js"
   },
   "license": "MIT",
   "dependencies": {
@@ -35,6 +35,7 @@
     "@babel/core": "^7.10.3",
     "@babel/preset-env": "^7.10.3",
     "@geo1088/eslint-config": "^3.1.0",
+    "@types/express-session": "^1.17.4",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.1.0",
     "css-loader": "^3.6.0",
@@ -45,6 +46,8 @@
     "sass": "^1.26.8",
     "sass-loader": "^8.0.2",
     "style-loader": "^1.2.1",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5",
     "vue-loader": "^15.9.6",
     "vue-template-compiler": "^2.6.11",
     "webpack": "^5.28.0",
