+++ src/bot/commands/.addfilter.js
