+++ src/bot/commands/.regulars.js
