+++ src/bot/commands/debug.js
