+++ src/bot/commands/misc/avatar.js
