+++ src/bot/commands/misc/joke.js
