+++ src/bot/commands/misc/pick.js
