+++ src/bot/commands/misc/roll.js
