+++ src/bot/commands/moderation/ban.js
