+++ src/bot/commands/moderation/filteradd.js
