+++ src/bot/commands/moderation/kick.js
