+++ src/bot/commands/moderation/note.js
