+++ src/bot/commands/ping.js
