+++ src/bot/commands/reload.js
