+++ src/bot/commands/setstatus.js
