+++ src/bot/commands/verify.js
