+++ src/bot/events/verification.js
