+++ src/bot/events/voiceChannelLeave.js
