+++ src/bot/util/formatting.js
