+++ src/web/middleware/responseHelpers.js
