+++ src/web/routes/discordAuth.js
@@ -8,7 +8,7 @@ const polka = require('polka');
 const fetch = require('node-fetch');
 const log = require('another-logger');
 
-const config = require('../../config');
+const config = require('../../../config');
 
 /** The redirect URI for Discord to send the user back to. */
 const discordRedirectURI = `${config.web.host}/auth/discord/callback`;
