+++ src/web/util.js
@@ -1,4 +1,4 @@
-const config = require('../config');
+const config = require('../../config');
 
 module.exports = {
 	asyncFilter: (arr, predicate) => Promise.all(arr.map(predicate)).then(results => arr.filter((_v, index) => results[index])),
